Add ability to customize section numbers
Add a hook to allow users to set a custom function to format section
numbers. Modify TOC parser and builders to invoke this.

This provides a mechanism for users to change the formatting of
automatic section numbers, for example, to use a different number style,
different separator between numbers for different levels, and/or to
insert additional text around the numbers.

Implements #4133.